+++ packages/gmana/lib/utilities.dart
@@ -1,7 +1,5 @@
 library;
 
-export 'design_system/spacing.dart';
-export 'math/wave_vertical_offset.dart';
 export 'utils/debouncer.dart';
 export 'utils/id_generator.dart';
 export 'utils/throttler.dart';